Add files via upload
This is all of the Reverse Engineered Gta 3 and Gta Vice City Source Code That was Modded by a Group of Hackers and Turned into a Mod that Remasters the Steam Versions of Gta 3 and Gta Vice City on Pc. it was Removed from Github due to a Copyright Lawsuit from Take Two interactive (The Company Who Owns Rockstar Games) and They Deleted all of it but i Downloaded The Entire Code Before it was Removed. So im Reuploading it Here and at a lot of other Places Because Fuck Rockstar and Fuck Take Two interactive.